% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postMCMCalterations.R
\name{compareModelsMeanPlots}
\alias{compareModelsMeanPlots}
\title{Compare models plot for fitted SSMs with the proposed framework for some specific parameters}
\usage{
compareModelsMeanPlots(
  models = list(),
  modelNames = NULL,
  fullModel = stateSpaceModel,
  nodes = c(),
  metrics = list(mean = TRUE, se = TRUE, confint = TRUE)
)
}
\arguments{
\item{models}{A list of MCMC results estimated from the sparta Updating functions in this paper}

\item{modelNames}{The names to be assigned to each model in the models parameter.}

\item{fullModel}{The full NIMBLE model used for running the updated model.}

\item{nodes}{The parameters we are interested in retrieving the ESS, efficiency and Monte Carlo Standard error for.}

\item{metrics}{list of logical responses for whether the plots of the following parameters should be returned: mean, se, confint.
If mean = TRUE, it plots the posterior mean of the nodes (parameters)
If se = TRUE, it plots the posterior mean plus or minus the standard error of the nodes (parameters).
If confint = TRUE, it plots the posterior mean of the nodes with the 95% credible intervals.
The MCse is estimated with mcmcse package and ESS was estimated with ggmcmc package in R}

\item{method}{The methods to be used in estimating the Monte Carlo standard error.
The choices of these should be the same as there is in the mcmcse package. The default chosen is batch means ("bm").}
}
\description{
Returns a dataframe with the times the models run, the effective sample size of each model, the efficiency of each model (where efficiency = ESS/times run) and the Monte Carlo standard error
}
\references{
Fernández-i-Marín, X. (2013). Using the ggmcmc Package.

Flegal, J. M., Hughes, J., Vats, D., Dai, N., Gupta, K., Maji, U., ... & Rcpp, L. (2017). Package ‘mcmcse’.
}
\seealso{
Other particle filtering methods: 
\code{\link{buildAuxiliaryFilterNew}},
\code{\link{buildAuxiliaryFilterUpdate}},
\code{\link{buildBootstrapFilterNew}},
\code{\link{buildBootstrapFilterUpdate}},
\code{\link{buildINLAmodel}},
\code{\link{compareModelsIndividualPars}()},
\code{\link{compareModelsPars}()},
\code{\link{compareModelsPlots}()}
}
\author{
Kwaku Peprah Adjei
}
\concept{particle filtering methods}
