% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postMCMCalterations.R
\name{compareModelsPars}
\alias{compareModelsPars}
\title{Compare models fitted with the proposed framework}
\usage{
compareModelsPars(
  models = list(),
  n.chains = NULL,
  modelNames = NULL,
  method = "bm",
  metrics = list(timeRun = TRUE, ESS = TRUE, efficiency = TRUE, MCse = TRUE)
)
}
\arguments{
\item{models}{A list of MCMC results estimated from the sparta Updating functions in this paper}

\item{n.chains}{The number of chains for the models. The n.chains should be equal for all models.}

\item{modelNames}{The names to be assigned to each model in the models parameter.
The default is NULL and the model names are created with the names 'Model 1', Model 2, ... Model N (where N = n.chains)}

\item{method}{The methods to be used in estimating the Monte Carlo standard error.
The choices of these should be the same as there is in the mcmcse package. The default chosen is batch means ("bm").}

\item{metrics}{list of logical responses for whether the following parameters should be returned: timeRun, ESS, efficiency, MCse.}
}
\description{
Returns a dataframe with the times the models run, the effective sample size of each model, the efficiency of each model (where efficiency = ESS/times run) and the Monte Carlo standard error
}
\references{
Flegal, J. M., Hughes, J., Vats, D., Dai, N., Gupta, K., Maji, U., ... & Rcpp, L. (2017). Package ‘mcmcse’.
}
\seealso{
Other particle filtering methods: 
\code{\link{buildAuxiliaryFilterNew}},
\code{\link{buildAuxiliaryFilterUpdate}},
\code{\link{buildBootstrapFilterNew}},
\code{\link{buildBootstrapFilterUpdate}},
\code{\link{buildINLAmodel}},
\code{\link{compareModelsIndividualPars}()},
\code{\link{compareModelsMeanPlots}()},
\code{\link{compareModelsPlots}()}
}
\author{
Kwaku Peprah Adjei
}
\concept{particle filtering methods}
