% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_road.R
\name{distGBIFdata}
\alias{distGBIFdata}
\title{Calculating distance to road from GBIF data}
\usage{
distGBIFdata(gbif_data, bioclim = TRUE)
}
\arguments{
\item{gbif_data, }{bioclim}
}
\value{
dataframe with distance to road a column
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
