% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaDataGenerating.R
\name{estimatingParsDataGenerating}
\alias{estimatingParsDataGenerating}
\title{INLABRU function for Citizen Science data}
\usage{
estimatingParsDataGenerating(omega)
}
\arguments{
\item{sppixels}{}
}
\value{
MCMC output
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
