% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_within_nimble.R
\name{fit.inla}
\alias{fit.inla}
\title{Fitting INLA within NIMBLE}
\usage{
fit.inla(x, y, beta, interInModel, family)
}
\arguments{
\item{data, }{code, family,n.iterations, n.chains, n.burnin}
}
\value{
MCMC output
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
