% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_with_inla_sim_inla.R
\name{fitINLAMissingValues}
\alias{fitINLAMissingValues}
\title{Conditional INLA for missing covariate}
\usage{
fitINLAMissingValues(missData, indexMiss, eta)
}
\arguments{
\item{data, }{code, family,n.iterations, n.chains, n.burnin}
}
\value{
dataframe
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
