% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGenerationSimulation.R
\name{generateCSData}
\alias{generateCSData}
\title{Generate Citizen Science data}
\usage{
generateCSData(input, domain = NULL, colmatrix = NULL)
}
\arguments{
\item{sppixels}{}
}
\value{
MCMC output
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
