% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaFunctionsDataGenerating.R
\name{im2rast}
\alias{im2rast}
\title{Convert image to raster}
\usage{
im2rast(im)
}
\arguments{
\item{sppixels}{}
}
\value{
MCMC output
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
