% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclass_nimble.R
\name{misclassFitModel}
\alias{misclassFitModel}
\title{Fit model using NIMBLE}
\usage{
misclassFitModel(
  simulations_all,
  type,
  model_selection,
  n.iterations,
  n.chains,
  n.burnin,
  parameters_to_monitor
)
}
\arguments{
\item{simulations_all, }{type, model_selection,
n.iterations, n.chains, n.burnin}
}
\value{
nimble output
}
\description{
This function takes the returned output of the misclassSetPars function
and generates multispecies misclassified data
}
