% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclass_sim.R
\name{misclassGenData}
\alias{misclassGenData}
\title{Generating data}
\usage{
misclassGenData(input, type, seed)
}
\arguments{
\item{input, }{type, seed}
}
\value{
Simulated data for the misclassification project
}
\description{
This function takes the returned output of the misclassSetPars function
and generates multispecies misclassified data
}
