% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclass_sim.R
\name{misclassSetPars}
\alias{misclassSetPars}
\title{Set the parameters}
\usage{
misclassSetPars(
  nspecies,
  nsites,
  nvisit,
  nvalidate,
  nreported,
  beta0,
  beta1,
  gamma0,
  gamma1,
  sigma_cov,
  sigma_class,
  sigma_corr,
  ncov
)
}
\arguments{
\item{nspecies, nsites, }{nvisit, nvalidate, nreported,
beta0, beta1, gamma0, gamma1, sigma_cov,
sigma_class, sigma_corr}
}
\value{
A list of parameters for simulating data
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
\examples{
 misclassSetPars(nspecies = 2, nsites = 1000,nvisit = 1,nvalidate = 100,
 nreported = 3, beta0 = c(1,0), beta1 = c(0,1), gamma0 = matrix(c(2, 1, 0,
                1, 3, 0), nrow = 2,
               ncol = 3, byrow = TRUE),
gamma1 = matrix(c(-1, 1,0 ,
                 -1, 2, 0), nrow = 2,
               ncol = 3, byrow = TRUE),
sigma_cov = 1,
sigma_class = 1,
sigma_corr = matrix(c(1,0.8,0.8,1), nrow = 2, byrow= T))
}
