% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_within_nimble.R
\name{nimbleINLA}
\alias{nimbleINLA}
\title{Compiling fit.inla in C+}
\usage{
nimbleINLA(x, y, beta, interInModel = 1, family = "gaussian")
}
\arguments{
\item{data, }{code, family,n.iterations, n.chains, n.burnin}
}
\value{
MCMC output
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
