% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaDataGenerating.R
\name{nimbleINLADataGenerating}
\alias{nimbleINLADataGenerating}
\title{Compiling function in C++}
\usage{
nimbleINLADataGenerating(omega)
}
\arguments{
\item{sppixels}{}
}
\value{
MCMC output
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
