% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_with_inla_sim_inla.R
\name{nimbleINLAMissingValues}
\alias{nimbleINLAMissingValues}
\title{Compiling fitINLAMissingValues in C+}
\usage{
nimbleINLAMissingValues(missData, indexMiss, eta)
}
\arguments{
\item{data, }{code, family,n.iterations, n.chains, n.burnin}
}
\value{
MCMC output
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
