% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaFunctionsDataGenerating.R
\name{rotate}
\alias{rotate}
\title{Fitting INLA within NIMBLE}
\usage{
rotate(x)
}
\arguments{
\item{data, }{code, family,n.iterations, n.chains, n.burnin}
}
\value{
MCMC output
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
