% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaFunctionsDataGenerating.R
\name{sppixels2raster}
\alias{sppixels2raster}
\title{Convert pixels to raster}
\usage{
sppixels2raster(sppixels)
}
\arguments{
\item{sppixels}{}
}
\value{
MCMC output
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
