% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_with_inla_sim_inla.R
\name{INLAWiNim}
\alias{INLAWiNim}
\title{Fitting INLA within NIMBLE}
\usage{
INLAWiNim(
  data,
  code,
  fam,
  modelData,
  modelConstants,
  modelInits,
  parametersToMonitor = c("beta", "sigma", "intercept"),
  mcmcControl = NULL,
  mcmcSamplerChange = FALSE,
  parametersForSamplerChange = NULL,
  newSampler = NULL,
  newSamplerControl = NULL,
  mcmcConfiguration = list(n.chains = 1, n.iterations = 10, n.burnin = 0, n.thin = 1,
    setSeed = TRUE, samples = TRUE, samplesAsCodaMCMC = TRUE, summary = TRUE, WAIC =
    TRUE)
)
}
\arguments{
\item{data, }{code, family,n.iterations, n.chains, n.burnin}
}
\value{
MCMC output
}
\description{
This function sets the paramters in the appropriate manner to be used by the
simulation function
}
